Separate code to operate process privileges from daemon
It's almost not related with daemon itself.
Multi process handling requires this feature, but requiring daemon.rb is too much.
Separating this code to a file makes code and dependency graph clean.
